+++ lib/spout/helpers/chart_types.rb
@@ -1,3 +1,4 @@
+require 'spout/helpers/array_statistics'
 require 'spout/helpers/table_formatting'
 
 module Spout
@@ -40,6 +41,16 @@ module Spout
         get_json(json['domain'], 'domain')
       end
 
+      def self.domain_array(variable_name)
+        variable_file = Dir.glob("variables/**/#{variable_name}.json").first
+        json = JSON.parse(File.read(variable_file)) rescue json = nil
+        if json
+          domain_json = get_domain(json)
+          domain_json ? domain_json.collect{|option_hash| [option_hash['display_name'], option_hash['value']]} : []
+        else
+          []
+        end
+      end
 
       def self.chart_arbitrary_choices_by_quartile(chart_type, subjects, json, method)
         # CHART TYPE IS THE QUARTILE VARIABLE
@@ -188,7 +199,7 @@ module Spout
 
       def self.chart_arbitrary_choices(chart_type, subjects, json, method)
         return unless chart_variable_json = get_variable(chart_type)
-        return unless chart_variable_domain = Spout::Commands::Graphs::domain_array(chart_type)
+        return unless chart_variable_domain = domain_array(chart_type)
         return unless domain_json = get_domain(json)
 
 
@@ -213,7 +224,7 @@ module Spout
 
       def self.chart_arbitrary(chart_type, subjects, json, method, visits)
         return unless chart_variable_json = get_variable(chart_type)
-        return unless chart_variable_domain = Spout::Commands::Graphs::domain_array(chart_type)
+        return unless chart_variable_domain = domain_array(chart_type)
         return chart_arbitrary_by_quartile(chart_type, subjects, json, method, visits) if ['numeric', 'integer'].include?(chart_variable_json['type'])
 
         return chart_arbitrary_choices(chart_type, subjects, json, method) if json['type'] == 'choices'
@@ -248,7 +259,7 @@ module Spout
 
       def self.table_arbitrary(chart_type, subjects, json, method, subtitle = nil)
         return unless chart_variable_json = get_variable(chart_type)
-        return unless chart_variable_domain = Spout::Commands::Graphs::domain_array(chart_type)
+        return unless chart_variable_domain = domain_array(chart_type)
         return table_arbitrary_by_quartile(chart_type, subjects, json, method, subtitle) if ['numeric', 'integer'].include?(chart_variable_json['type'])
         return table_arbitrary_choices(chart_type, subjects, json, method, subtitle) if json['type'] == 'choices'
 
@@ -284,7 +295,7 @@ module Spout
 
       def self.table_arbitrary_choices(chart_type, subjects, json, method, subtitle)
         return unless chart_variable_json = get_variable(chart_type)
-        return unless chart_variable_domain = Spout::Commands::Graphs::domain_array(chart_type)
+        return unless chart_variable_domain = domain_array(chart_type)
         return unless domain_json = get_domain(json)
 
         headers = [
@@ -330,7 +341,7 @@ module Spout
       def self.chart_histogram_choices(chart_type, subjects, json, method)
         return unless domain_json = get_domain(json)
         return unless chart_variable_json = get_variable(chart_type)
-        return unless chart_variable_domain = Spout::Commands::Graphs::domain_array(chart_type)
+        return unless chart_variable_domain = domain_array(chart_type)
 
 
         title = "#{json['display_name']}"
@@ -360,7 +371,7 @@ module Spout
       def self.chart_histogram(chart_type, subjects, json, method)
         return chart_histogram_choices(chart_type, subjects, json, method) if json['type'] == 'choices'
         return unless chart_variable_json = get_variable(chart_type)
-        return unless chart_variable_domain = Spout::Commands::Graphs::domain_array(chart_type)
+        return unless chart_variable_domain = domain_array(chart_type)
 
         title = "#{json['display_name']}"
         subtitle = "By Visit"